Cairo GAL: reduce buffer allocation sizes.
cairo_format_stride_for_width() result is in bytes.
Also removes unneeded extra width.


(cherry picked from commit d359fb8ab90937d3a5b4dc4cfbabf7dc04da2952)

+++ include/gal/cairo/cairo_compositor.h
@@ -126,7 +126,7 @@ protected:
         return m_buffers.size();
     }
 
-    typedef uint32_t* BitmapPtr;
+    typedef uint8_t* BitmapPtr;
     struct CAIRO_BUFFER
     {
         cairo_t*            context;        ///< Main texture handle
